EntreesREADME.md
A program that tracks how much of each entree is sold and what the total is earned from each individual entree, as well as the total earned from ALL entrees added together for the day. The entrees on the menu are spaghetti with meatballs ($10.00), rib eye steak and potatoes ($25.00), cobb salad ($12.00), Red Snapper in butter sauce ($30.00) and Margherita Pizza ($15.95). Ask for user input of how many of each entrees was sold each day, Monday through Friday. Use this program to tally each day's total from Monday to Friday and then add each day's total to get the final amount earned over the week (Monday to Friday).